One first grade of testability can be derived: code is automatically testable if it has no dependency to the UI.
In this example we had dependencies to the UI in lines 9 to 10 and line 12.

To separate logic that is easy to test from the difficult one, we move them to separate functions.

But the overall testability of the behaviour has not improved, yet. Likewise we cannot explicitly test the domain logic (the counting of words) because it is caught between die UI method calls. This is a contradiction to the "Single Level of Abstraction" principle (http://principles-wiki.net/principles:single_level_of_abstraction).

Lines 9 and 12 have now abstraction level 3 (logic behind a function name), lines 10 and 11 have abstraction level 2 ( logic named with a variable name).