The test TestCountingOfWords lead to the refactoring of the method SplitByWhitespace() which now removes empty entries.
From here on I will mark test methods via comments as an acceptance test or as a scaffolding test. Scaffolding test can be deleted since they are not necessary anymore if the acceptance test yields a correct result.

Our acceptance tests are until now TestDomainLogic() and for the sake of completeness TestDomainLogicWithMultipleWhitespaces().