We simply read the lines in our domain logic that counts the words after splitting the text.
The new test passes but all others fail (except TestSplittingOfText())! The other tests use the stop words aspect, too, but they are not designed to consider it.

The failing of the other tests is a sign of dependency, which is hidden deeply  in the domain's aspect "counting of words".

Therefore not only a dependency to an UI is a test impediment. Every dependency to a ressource degrades testability.