The loading of stopwords itself has different aspects, too. The current design impedes the test of e.g.
* missing stopwords file
* empty lines in the file

Since LoadStopwords() uses a constant (the file name) we learn that testabilty degrades with dependencies to data in constants or variables.

To make the file name exchangeable we introduce the class StopwordsProvider that is responsible for loading the stopwords (good for SRP principle, too :)). It can get a file name via a constructor but does not need so.

We also refactor TestLoadStopwords(). It uses now the new class StopwordsProvider and tries to load a different file.