Why should the method WordsCounter.Count() should have a functional dependency on the method Load()? Why should the WordsCounter be responsible for loading the data, too?
That is another violation of SRP and functional dependencies are problematic for testability.

The problem is basically solved by removing the functional dependency, i. e. the WordsCounter must not know be responsible for loading the stopwords anymore. It only needs the stopwords.

Therefore the WordsCounter gets its stopwords via the constructor.

It doesn't matter if the loading is now used via Dependency Inversion.. The injection of the provider creates a load to build a mock.

The MockstopwordsProvider can be deleted, also the tests

* TestCountingOfWords()
* TestCountingOfWordsWithStopwords()
* TestLoadStopwords()
* TestMain()

They refer to code we have refactored or will be soon. The scaffolding test TestSplittingOfWords() is not needed anymore, too.

Instead we write some new tests for the WordsCounter which pass.