At last we merge the two methods CountWords() into one.
Since the result is an integration method (IOSP, https://ccd-school.de/tag/iosp/) and does not contain any logic we must not test it. We can set its accessability to private and remove all tests.

Of course the tests

* TestDomainLogic()
* TestDomainLogicWithMultipleWhitespaces()

now fail but we can delete them since all the functionality is or was tested in other tests.

If we need to test it again we can make it public again.